Integrate lipstick with temp test views
Integrated lipstick with temporary test view and controllers. Some
auth measures were temporarily disabled so that the views could be
be seen in this PR. This will be resolved in the a future PR.

There was an issue upstream which meant the `torba-rails` gem had to
be used rather than the `torba` gem.