Support for multi-value attributes
- Every attribute returns an array of values now
- The attribute values are sourced from the snapshot values, if no
snapshot value exists, it provides a nil value to the array
- This array of values is run through the validation sequence.  We use
the most important (determined by state order) attribute state to
determine which group the parent federation attribute gets bundled
into.  This means that a federation attribute with an invalid attribute
value is placed higher on the list even if the other attribute values
are good.