Fixes filtering problem on RailsAdmin virtual fields
It makes sense that these weren’t working.  Virtual (associated fields)
only allow searching on the foreign key, and have to be told if that is
a different field.

CategoryAttributes listing FederationAttributes are unable to grab the
primary_alias name, as it is highly delegated field.  Using
internal_alias is not perfect but close enough for most cases.