Identify `Subject` by generic unique identifier
Previously we explicitly only identified a `Subject` by the value
transported to the application as `eduPersonTargetdID` within a SAML
attribute assertion.

This is no longer the case, identifiers may be transported as a
SAML Persistent NameID or eduPersonTargetdID, the former being
preferred.

With this change Subjects are identified by Persistent NameID if present
in the assertion and fall back to eduPersonTargetdID when necessary.