Regenerate brakeman cfg, add HTTP verb confusion
This rule is quite interesting, see https://github.com/presidentbeef/brakeman/pull/1524,
https://github.com/presidentbeef/brakeman/issues/1432. However it does not
make sense to try to adhere to it, this was the warning:

```
Confidence: Weak
Category: HTTP Verb Confusion
Check: VerbConfusion
Message: Potential HTTP verb confusion. `HEAD` is routed like `GET` but
         `request.get?` will return `false`
Code: session[:return_url] = request.url if request.get?
File: app/controllers/application_controller.rb
Line: 56
```

The only way to appease this rule is to add an "elseif" on our use of
`request.get?`. We can't reasonably do this in this case. So this rule
can be ignored (and for anything else, as there will be no further code).